orm: improve tables race test
